// src/shared/enum/dto.enum.ts
export enum EAuthSignInDTO {
    loginEmpty = "Логин не может быть пустым.",
    IsEmailOrPhone = "Логин должен быть почтой или номером телефона!",
    passwordEmpty = "Пароль не может быть пустым.",
    strongPassword = "Пароль слишком слабый. Используйте минимум 6 символов, включая заглавные, строчные буквы, цифры и спецсимволы.",
    email = 'Введите корректный email.',
    tel= 'Введите корректный номер телефона.',
    telEmpty = 'Телефон обязателен!',
    emailEmpty = 'Почта обязателна!',
    lengthVerifyCode = 'Длина кода должна быть равной 6.',
    key = 'Что-то не так... Попробуйте заново.',
    empty = 'Параметр не может быть пустым!',
    IsNotProfanity='Маты не допустимы в ФИО!',
    MaxLength = 'Вы превысили максимально допустимую длину в параметре!'
}
export enum EUserUpdDataDTO {
    isNotProfanity = 'ФИО не может содержать брань или нецензурную лексику!',
    isPhone= 'Введите корректный номер телефона.',
}